+++ src/main/java/lotto/Application.java
@@ -1,6 +1,9 @@
 package lotto;
 
 
+import lotto.model.Budget;
+import lotto.view.InputView;
+
 public class Application {
 
     public static void main(String[] args) {
