+++ src/main/java/lotto/domain/Lotto.java
@@ -1,10 +1,11 @@
-package lotto;
+package lotto.domain;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
+import lotto.common.ErrorMessage;
 
 public class Lotto {
     private static final int MIN_NUMBER = 1;
