+++ src/main/java/lotto/view/InputView.java
@@ -1,6 +1,7 @@
-package lotto;
+package lotto.view;
 
 import camp.nextstep.edu.missionutils.Console;
+import lotto.model.Budget;
 
 public class InputView {
     private static final String INPUT_BUDGET_INFO = "구입금액을 입력해 주세요.";
