+++ src/test/java/lotto/domain/LottiesTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import static org.junit.jupiter.api.Assertions.*;
