+++ src/test/java/lotto/view/InputViewTest.java
@@ -1,0 +1,20 @@
+package lotto.view;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+class InputViewTest {
+
+    @DisplayName("입력받은 구입 금액의 파싱이 불가능하면 예외를 일으킨다.")
+    @Test
+    void ERROR_BUDGET_NUMBER_FORMAT() {
+
+    }
+
+    @DisplayName("구입 금액 입력에서 예외가 발생하면 다시 입력받는다.")
+    @Test
+    void RETRY_INPUT_WHEN_BUDGET_ERROR() {
+
+    }
+
+}
